+++ bdtopo_v2/src/main/groovy/org/orbisgis/geoclimate/bdtopo_v2/BDTopo_V2.groovy
@@ -13,8 +13,8 @@ abstract class BDTopo_V2 extends GroovyProcessFactory {
     public static def logger = LoggerFactory.getLogger(BDTopo_V2.class)
 
     public static WorkflowBDTopo_V2 = new WorkflowBDTopo_V2()
-    public static InputDataLoading = new InputDataLoading2()
-    public static InputDataFormatting = new InputDataFormatting2()
+    public static InputDataLoading = new InputDataLoading()
+    public static InputDataFormatting = new InputDataFormatting()
 
 
     static def uuid = { UUID.randomUUID().toString().replaceAll("-", "_") }
